pull user query into a hook
